feat(playground): submit form on Enter key
