+++ src/components/Playground/OAPlaygroundResponse.vue
@@ -1,116 +1,51 @@
 <script setup lang="ts">
-import { computed, defineProps } from 'vue'
-import OACodeBlock from '../Common/OACodeBlock.vue'
+import { defineProps } from 'vue'
+import { Badge } from '../ui/badge'
+import OAPlaygroundResponseContent from './OAPlaygroundResponseContent.vue'
 
-const { response } = defineProps({
+const { response, loading } = defineProps({
   response: {
-    type: Object,
-    required: true,
+    type: null,
+    required: false,
+  },
+  loading: {
+    type: Boolean,
+    default: false,
   },
 })
-
-const isJson = /json/i.test(response.type)
-const isXml = /xml/i.test(response.type)
-const isHtml = /text\/html/i.test(response.type)
-const isPlainText = /text\/plain/i.test(response.type)
-const isCsv = /text\/csv/i.test(response.type)
-const isImage = /^image\//i.test(response.type)
-const isAudio = /^audio\//i.test(response.type)
-const isDownloadable = /^application\/octet-stream/i.test(response.type)
-  || (
-    response
-    && response.headers
-    && (
-      /attachment/i.test(response.headers['Content-Disposition'])
-      || /attachment/i.test(response.headers['content-disposition'])
-      || /download/i.test(response.headers['Content-Disposition'])
-      || /download/i.test(response.headers['content-disposition'])
-    )
-  )
-
-const lang = computed(() => {
-  if (isJson) {
-    return 'json'
-  }
-  if (isXml) {
-    return 'xml'
-  }
-  if (isHtml) {
-    return 'html'
-  }
-  if (isPlainText) {
-    return 'plaintext'
-  }
-  if (isCsv) {
-    return 'csv'
-  }
-  return 'plaintext'
-})
-
-const label = computed(() => {
-  if (isJson) {
-    return 'JSON'
-  }
-  if (isXml) {
-    return 'XML'
-  }
-  if (isHtml) {
-    return 'HTML'
-  }
-  if (isPlainText) {
-    return 'Plain Text'
-  }
-  if (isCsv) {
-    return 'CSV'
-  }
-  return 'Plain Text'
-})
-
-const url = computed(() => {
-  if (isAudio) {
-    try {
-      return window.URL ? URL.createObjectURL(response.body) : ''
-    } catch { }
-  }
-
-  return ''
-})
-
-const disableHtmlTransform = computed(() => response.body && JSON.stringify(response.body).length > 1000)
-
-function downloadBlob(blob: Blob, fileName: string) {
-  const url = window.URL.createObjectURL(blob)
-  const link = document.createElement('a')
-  link.href = url
-  link.download = fileName
-  link.click()
-  window.URL.revokeObjectURL(url)
-}
 </script>
 
 <template>
-  <div>
-    <OACodeBlock
-      v-if="isJson || isXml || isHtml || isPlainText || isCsv"
-      :code="response.body"
-      :lang="lang"
-      :label="label"
-      :disable-html-transform="disableHtmlTransform"
-      class="!m-0"
-    />
-    <img v-else-if="isImage" :src="response.body" alt="Response Image">
-    <audio v-else-if="isAudio" controls class="w-full mt-2">
-      <source :src="url" :type="response.type">
-      {{ $t('Your browser does not support the audio element.') }}
-    </audio>
-    <div v-else-if="isDownloadable">
-      <button @click="downloadBlob(response.body, 'response_file')">
-        {{ $t('Download file') }}
-      </button>
+  <details v-if="response || loading" open>
+    <summary class="!my-0 text-lg font-bold cursor-pointer">
+      <div class="inline-flex items-center gap-2 w-[calc(100%-24px)]">
+        <span>{{ loading ? $t('Loading...') : $t('Response') }}</span>
+
+        <span class="flex-1" />
+
+        <Badge
+          v-if="response && response.status"
+          variant="plain"
+          class="rounded py-1.5"
+          :class="{
+            'bg-green-100 text-green-800 dark:bg-green-800 dark:text-green-100': String(response.status).startsWith('2'),
+            'bg-yellow-100 text-yellow-800 dark:bg-yellow-800 dark:text-yellow-100': String(response.status).startsWith('3'),
+            'bg-red-100 text-red-800 dark:bg-red-800 dark:text-red-100': String(response.status).startsWith('4') || String(response.status).startsWith('5'),
+          }"
+        >
+          {{ response.status }}
+        </Badge>
+      </div>
+    </summary>
+
+    <div class="flex flex-col gap-2">
+      <div v-if="response" class="text-sm text-gray-500">
+        {{ $t('Response time') }}: {{ loading ? $t('Loading...') : `${response.time}ms` }}
+      </div>
+
+      <div class="flex flex-col max-h-96 overflow-y-auto">
+        <OAPlaygroundResponseContent v-if="response" :response="response" />
+      </div>
     </div>
-    <div v-else>
-      <p>{{ $t('Unrecognized response type. Raw content:') }}</p>
-      <pre class="whitespace-pre-wrap">{{ response.body }}</pre>
-    </div>
-  </div>
+  </details>
 </template>
